move implementation of ProfileDict and EventedDict to utils module
