+++ package/PartSeg/common_backend/base_settings.py
@@ -36,10 +36,11 @@ from PartSegCore import register
 from PartSegCore.color_image import default_colormap_dict, default_label_dict
 from PartSegCore.color_image.base_colors import starting_colors
 from PartSegCore.io_utils import load_metadata_base
-from PartSegCore.json_hooks import PartSegEncoder, ProfileDict, check_loaded_dict
+from PartSegCore.json_hooks import PartSegEncoder
 from PartSegCore.project_info import AdditionalLayerDescription, HistoryElement, ProjectInfoBase
 from PartSegCore.roi_info import ROIInfo
 from PartSegCore.segmentation.algorithm_base import ROIExtractionResult
+from PartSegCore.utils import ProfileDict, check_loaded_dict
 from PartSegImage import Image
 
 if hasattr(napari.utils.theme, "get_theme"):
