+++ package/tests/test_PartSegCore/test_json_hooks.py
@@ -3,7 +3,6 @@
 import json
 from dataclasses import dataclass
 from enum import Enum
-from unittest.mock import MagicMock
 
 import numpy as np
 import pytest
@@ -14,14 +13,8 @@ from pydantic import BaseModel
 from PartSegCore._old_json_hooks import ProfileEncoder, profile_hook
 from PartSegCore.class_register import class_to_str, register_class, rename_key
 from PartSegCore.image_operations import RadiusType
-from PartSegCore.json_hooks import (
-    EventedDict,
-    PartSegEncoder,
-    ProfileDict,
-    add_class_info,
-    partseg_object_hook,
-    recursive_update_dict,
-)
+from PartSegCore.json_hooks import PartSegEncoder, add_class_info, partseg_object_hook
+from PartSegCore.utils import ProfileDict
 
 
 @dataclass
@@ -33,7 +26,7 @@ class SampleDataclass:
 class SamplePydantic(BaseModel):
     sample_int: int
     sample_str: str
-    sample_datacls: SampleDataclass
+    sample_dataclass: SampleDataclass
 
 
 class SampleAsDict:
@@ -43,205 +36,6 @@ class SampleAsDict:
 
     def as_dict(self):
         return {"value1": self.value1, "value2": self.value2}
-
-
-def test_recursive_update_dict_basic():
-    dict1 = {"a": 1, "b": 2}
-    dict2 = {"b": 3, "c": 4}
-    dict1_copy = dict1.copy()
-    dict1.update(dict2)
-    recursive_update_dict(dict1_copy, dict2)
-    assert dict1 == dict1_copy
-
-
-def test_recursive_update_dict():
-    dict1 = {"a": {"k": 1, "l": 2}, "b": {"k": 1, "l": 2}}
-    dict2 = {"a": {"m": 3, "l": 4}, "b": 3, "c": 4}
-    recursive_update_dict(dict1, dict2)
-    assert dict1 == {"a": {"k": 1, "l": 4, "m": 3}, "b": 3, "c": 4}
-
-
-class TestEventedDict:
-    def test_simple_add(self):
-        receiver = MagicMock()
-
-        dkt = EventedDict()
-        dkt.setted.connect(receiver.set)
-        dkt.deleted.connect(receiver.delete)
-        dkt["a"] = 1
-        assert dkt["a"] == 1
-        assert receiver.set.call_count == 1
-        assert "'a': 1" in str(dkt)
-        assert "'a': 1" in repr(dkt)
-        dkt["a"] = 2
-        assert dkt["a"] == 2
-        assert receiver.set.call_count == 2
-        assert len(dkt) == 1
-        del dkt["a"]
-        assert receiver.set.call_count == 2
-        assert receiver.delete.call_count == 1
-        assert len(dkt) == 0
-
-    def test_simple_add_remove(self):
-        callback_list = []
-
-        def callback_add():
-            callback_list.append(1)
-
-        def callback_delete():
-            callback_list.append(2)
-
-        dkt = EventedDict()
-        dkt.setted.connect(callback_add)
-        dkt.deleted.connect(callback_delete)
-
-        dkt[1] = 1
-        dkt[2] = 1
-        assert len(dkt) == 2
-        assert callback_list == [1, 1]
-        del dkt[1]
-        assert len(dkt) == 1
-        assert callback_list == [1, 1, 2]
-
-    def test_nested_evented(self):
-        dkt = EventedDict(bar={"foo": {"baz": 1}})
-        assert isinstance(dkt["bar"], EventedDict)
-        assert isinstance(dkt["bar"]["foo"], EventedDict)
-        assert dkt["bar"]["foo"]["baz"] == 1
-
-        dkt["baz"] = {"bar": {"foo": 1}}
-        assert isinstance(dkt["baz"], EventedDict)
-        assert isinstance(dkt["baz"]["bar"], EventedDict)
-        assert dkt["baz"]["bar"]["foo"] == 1
-
-    def test_serialize(self, tmp_path):
-        dkt = EventedDict(
-            **{"a": {"b": {"c": 1, "d": 2, "e": 3}, "f": 1}, "g": {"h": {"i": 1, "j": 2}, "k": [6, 7, 8]}}
-        )
-        with (tmp_path / "test_dict.json").open("w") as f_p:
-            json.dump(dkt, f_p, cls=ProfileEncoder)
-
-        with (tmp_path / "test_dict.json").open("r") as f_p:
-            dkt2 = json.load(f_p, object_hook=profile_hook)
-
-        assert isinstance(dkt2, EventedDict)
-        assert isinstance(dkt2["a"], EventedDict)
-        assert dkt["g"]["k"] == [6, 7, 8]
-
-    def test_signal_names(self):
-        receiver = MagicMock()
-        dkt = EventedDict(baz={"foo": 1})
-        dkt.setted.connect(receiver.set)
-        dkt.deleted.connect(receiver.deleted)
-        dkt["foo"] = 1
-        assert receiver.set.call_count == 1
-        receiver.set.assert_called_with("foo")
-        dkt["bar"] = EventedDict()
-        assert receiver.set.call_count == 2
-        receiver.set.assert_called_with("bar")
-        dkt["bar"]["baz"] = 1
-        assert receiver.set.call_count == 3
-        receiver.set.assert_called_with("bar.baz")
-        dkt["baz"]["foo"] = 2
-        assert receiver.set.call_count == 4
-        receiver.set.assert_called_with("baz.foo")
-
-        del dkt["bar"]["baz"]
-        assert receiver.deleted.call_count == 1
-        receiver.deleted.assert_called_with("bar.baz")
-        del dkt["bar"]
-        assert receiver.deleted.call_count == 2
-        receiver.deleted.assert_called_with("bar")
-
-    def test_propagate_signal(self):
-        receiver = MagicMock()
-        dkt = EventedDict(baz={"foo": 1})
-        dkt.setted.connect(receiver.set)
-        dkt.deleted.connect(receiver.deleted)
-        dkt["baz"].base_key = ""
-        dkt["baz"]["foo"] = 2
-        receiver.set.assert_called_with("foo")
-        receiver.set.assert_called_once()
-        del dkt["baz"]["foo"]
-        receiver.deleted.assert_called_with("foo")
-        receiver.deleted.assert_called_once()
-
-
-class TestProfileDict:
-    def test_simple(self):
-        dkt = ProfileDict()
-        dkt.set("a.b.c", 1)
-        dkt.set("a.b.a", 2)
-        assert dkt.get("a.b.c") == 1
-        with pytest.raises(KeyError):
-            dkt.get("a.b.d")
-        dkt.get("a.b.d", 3)
-        assert dkt.get("a.b.d") == 3
-        assert dkt.get("a.b") == {"a": 2, "c": 1, "d": 3}
-        with pytest.raises(TypeError):
-            dkt.set("a.b.c.d", 3)
-
-    def test_update(self):
-        dkt = ProfileDict()
-        dkt.update(a=1, b=2, c=3)
-        assert dkt.my_dict == {"a": 1, "b": 2, "c": 3}
-        dkt2 = ProfileDict()
-        dkt2.update(c=4, d={"a": 2, "e": 7})
-        assert dkt2.get("d.e") == 7
-        dkt.update(dkt2)
-        assert dkt.get("d.e") == 7
-        assert dkt.get("c") == 4
-        dkt.update({"g": 1, "h": 4})
-        assert dkt.get("g") == 1
-        dkt.update({"w": 1, "z": 4}, w=3)
-        assert dkt.get("w") == 3
-        assert dkt.verify_data()
-        assert dkt.filter_data() == []
-        dkt.set("e.h.l", {"aaa": 1, "__error__": True})
-        assert not dkt.verify_data()
-        assert dkt.filter_data() == ["e.h"]
-
-    def test_serialize(self, tmp_path):
-        dkt = ProfileDict()
-        dkt.set("a.b.c", 1)
-        dkt.set("a.b.a", 2)
-        with open(tmp_path / "test.json", "w") as f_p:
-            json.dump(dkt, f_p, cls=ProfileEncoder)
-        with open(tmp_path / "test.json") as f_p:
-            dkt2 = json.load(f_p, object_hook=profile_hook)
-
-        assert dkt.my_dict == dkt2.my_dict
-
-    def test_callback(self):
-        def dummy_call():
-            receiver.dummy()
-
-        receiver = MagicMock()
-
-        dkt = ProfileDict()
-        dkt.connect("", receiver.empty)
-        dkt.connect("", dummy_call)
-        dkt.connect("b", receiver.b)
-        dkt.connect(["d", "c"], receiver.dc)
-
-        dkt.set("test.a", 1)
-        assert receiver.empty.call_count == 1
-        assert receiver.dummy.call_count == 1
-        receiver.empty.assert_called_with("a")
-        receiver.dummy.assert_called_with()
-        dkt.set("test.a", 1)
-        assert receiver.empty.call_count == 1
-        receiver.b.assert_not_called()
-        dkt.set("test2.a", 1)
-        assert receiver.empty.call_count == 2
-        receiver.b.assert_not_called()
-        dkt.set(["test", "b"], 1)
-        assert receiver.empty.call_count == 3
-        assert receiver.b.call_count == 1
-        dkt.set("test.d.c", 1)
-        receiver.dc.assert_called_once()
-        dkt.set("test.a", 2)
-        assert receiver.empty.call_count == 5
 
 
 def test_profile_hook_colormap_load(bundle_test_dir):
@@ -324,7 +118,7 @@ class TestPartSegEncoder:
     def test_pydantic_serialize(self, tmp_path):
         data = {
             "color1": Colormap(colors=[[0, 0, 0], [0, 0, 0]], controls=[0, 1]),
-            "other": SamplePydantic(sample_int=1, sample_str="text", sample_datacls=SampleDataclass(1, "text")),
+            "other": SamplePydantic(sample_int=1, sample_str="text", sample_dataclass=SampleDataclass(1, "text")),
         }
         with (tmp_path / "test.json").open("w") as f_p:
             json.dump(data, f_p, cls=PartSegEncoder)
@@ -332,7 +126,7 @@ class TestPartSegEncoder:
             data2 = json.load(f_p, object_hook=partseg_object_hook)
         assert data2["color1"] == Colormap(colors=[[0, 0, 0], [0, 0, 0]], controls=[0, 1])
         assert isinstance(data2["other"], SamplePydantic)
-        assert isinstance(data2["other"].sample_datacls, SampleDataclass)
+        assert isinstance(data2["other"].sample_dataclass, SampleDataclass)
 
     def test_numpy_serialize(self, tmp_path):
         data = {"arr": np.arange(10), "f": np.float32(0.1), "i": np.int16(1000)}
