+++ package/tests/test_PartSegCore/test_segmentation_proceses.py
@@ -8,8 +8,8 @@ import pytest
 from PartSegCore.algorithm_describe_base import ROIExtractionProfile
 from PartSegCore.analysis.algorithm_description import AnalysisAlgorithmSelection
 from PartSegCore.analysis.load_functions import UpdateLoadedMetadataAnalysis
-from PartSegCore.json_hooks import check_loaded_dict
 from PartSegCore.segmentation.algorithm_base import ROIExtractionAlgorithm
+from PartSegCore.utils import check_loaded_dict
 from PartSegImage import TiffImageReader
